requirements.txt
streamlit 
gspread 
oauth2client 
torch 
torchvision 
transformers 
scikit-learn 
pillow
pandas
numpy==1.26.4